Rename AsyncUtil to ViewUtil and move Amazon load method to Bidders.kt
