+++ dynamicprice/android/src/androidMain/kotlin/Bidders.kt
@@ -5,14 +5,19 @@ import com.adsbynimbus.lineitem.DEFAULT_BANNER
 import com.adsbynimbus.lineitem.applyDynamicPrice
 import com.adsbynimbus.request.NimbusRequest
 import com.adsbynimbus.request.NimbusResponse
+import com.amazon.device.ads.AdError
+import com.amazon.device.ads.DTBAdCallback
 import com.amazon.device.ads.DTBAdRequest
 import com.amazon.device.ads.DTBAdResponse
 import com.amazon.device.ads.DTBAdUtil
 import com.google.android.gms.ads.admanager.AdManagerAdRequest
 import kotlinx.coroutines.async
 import kotlinx.coroutines.supervisorScope
+import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlinx.coroutines.withTimeout
 import kotlinx.coroutines.yield
+import kotlin.coroutines.resume
+import kotlin.coroutines.resumeWithException
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
 
@@ -57,6 +62,20 @@ value class NimbusBidder(private val adRequest: NimbusRequest) : Bidder<NimbusRe
 @JvmInline
 value class ApsBidder(private val adRequest: () -> DTBAdRequest) : Bidder<DTBAdResponse> {
     override suspend fun fetchBid(): Bid<DTBAdResponse> = Bid(adRequest().loadAsync())
+}
+
+/** Loads an APS ad using a coroutine */
+suspend fun DTBAdRequest.loadAsync(): DTBAdResponse = suspendCancellableCoroutine { coroutine ->
+    val callback = object : DTBAdCallback {
+        override fun onFailure(p0: AdError) {
+            if (coroutine.isActive) coroutine.resumeWithException(RuntimeException(p0.message))
+        }
+
+        override fun onSuccess(p0: DTBAdResponse) {
+            if (coroutine.isActive) coroutine.resume(p0)
+        }
+    }
+    loadAd(callback)
 }
 
 /** Applies targeting values from a Bid to an AdManagerAdRequest.Builder */
